+++ src/builtin-applications/export/StaticFiles.js
@@ -106,7 +106,7 @@ export class StaticFiles extends BaseApplication {
       this.#notFoundOptions = {
         ...(StaticFiles.#SEND_OPTIONS),
         body:        await fs.readFile(notFoundPath),
-        contentType: MimeTypes.typeFromExtension(notFoundPath)
+        contentType: MimeTypes.typeFromPathExtension(notFoundPath)
       };
     }
   }