+++ src/net-protocol/export/Request.js
@@ -46,15 +46,16 @@ import { WranglerContext } from '#x/WranglerContext';
  * * `{Buffer|string|null} body` -- Body content to include. If passed as a
  *   `string`, the actual body gets encoded as UTF-8, and the `contentType` is
  *   adjusted if necessary to indicate that fact. This property is only
- *   available on methods that don't take a separate `body` argument.
+ *   available on methods that (a) could conceivably send a body, and (b) don't
+ *   have any other way to specify a body (e.g. through other arguments).
  * * `{?string} bodyExtra` -- Extra body content to include when wanting to
  *   _mostly_ use a default body. This is only available on {@link
  *   #sendMetaResponse}.
  * * `{?string} contentType` -- Content type of the body, in the form expected
- *   by {@link MimeTypes}. Required if `body` is present and ignored in all
- *   other cases. If this value starts with `text/` and/or the `body` is passed
- *   as a string, then the actual `Content-Type` header will indicate a charset
- *   of `utf-8`.
+ *   by {@link MimeTypes#typeFromExtensionOrType}. Required if `body` is present
+ *   and ignored in all other cases. If this value starts with `text/` and/or
+ *   the `body` is passed as a string, then the actual `Content-Type` header
+ *   will indicate a charset of `utf-8`.
  * * `{?Cookies} cookies` -- Cookies to include (via `Set-Cookie` headers) in
  *   the response, if any.
  * * `{?object} headers` -- Extra headers to include in the response, if any.
@@ -63,8 +64,10 @@ import { WranglerContext } from '#x/WranglerContext';
  *   subclass, {@link HttpHeaders}.
  * * `{?number} maxAgeMsec` -- Value to send back in the `max-age` property of
  *   the `Cache-Control` response header. Defaults to `0`.
- * * `{?number} statusCode` -- Response status code. This is only available on
- *   {@link #sendRedirect}.
+ * * `{?number} status` -- Response status code. This is only available on a
+ *   couple methods (as noted in their documentation) which (a) don't strongly
+ *   imply a status (or possible set thereof), and (b) don't have any other way
+ *   to specify a status (e.g. through other arguments).
  */
 export class Request {
   /**
@@ -521,9 +524,8 @@ export class Request {
    *   not sending a body at all). If passed as a `string`, it is encoded as
    *   UTF-8, and the content type of the response will list that as the
    *   charset.
-   * @param {string} contentType Content type for the body. If this value starts
-   *   with `text/` and/or the `body` is passed as a string, then the actual
-   *   `Content-Type` header will indicate a charset of `utf-8`.
+   * @param {string} contentType Content type for the body, in a form as
+   *   described by this class's header comment.
    * @param {object} [options] Options to control response behavior. See class
    *   header comment for more details. Header-related options are only used
    *   for non-error responses.
@@ -598,7 +600,7 @@ export class Request {
 
     const headers = this.#makeResponseHeaders('cacheable', options, {
       'content-type': () => {
-        return MimeTypes.typeFromExtension(path);
+        return MimeTypes.typeFromPathExtension(path);
       }
     });
 
@@ -779,8 +781,8 @@ export class Request {
    * @param {string} target Possibly-relative target URL.
    * @param {?object} [options] Options to control response behavior. See class
    *   header comment for more details.
-   * @param {?number} [options.statusCode] The status code to report. Defaults
-   *   to `302` ("Found").
+   * @param {?number} [options.status] The status code to report. Defaults to
+   *   `302` ("Found").
    * @returns {boolean} `true` when the response is completed.
    */
   async sendRedirect(target, options = null) {
@@ -807,8 +809,8 @@ export class Request {
    *
    * @param {?object} [options] Options to control response behavior. See class
    *   header comment for more details.
-   * @param {?number} [options.statusCode] The status code to report. Defaults
-   *   to `302` ("Found").
+   * @param {?number} [options.status] The status code to report. Defaults to
+   *   `302` ("Found").
    * @returns {boolean} `true` when the response is completed.
    */
   async sendRedirectBack(options = null) {